did more modularization. added dom utils js file and added to it. fixed circular dependency by making sure nothing was ever imorting from main.js
